Put the piece in a automatic position
Now, the human player don't need to locate the piece
in the corect position in the board. When is selected,
the game check if is right and put it on the table.

The logic used to put the piece on the table need improvements,
is confused when is at the border of the table,
and the positions selected are not optimal.

This change makes the game state LOCATE_PIECE redundant,
because the location is automatic.