Chnages to use more screen state
In the XO, with the actual pieces size, there are not
too much space to put the pieces.
The previous strategy required not use the tiles in the border of the board
That is too much space wasted.
To try to make a better use, changed the algorithm
to put pieces until the border of the board
and show the pieces of the automati player (flipped) smaller.