Use jpg images for the background
These images re smaller than the pngs, but need a intermediate step
to use them as paterns. We use a big image with all the screen size
right now, then we have tw images, one for the screen horizontal and
another for the screen vertical.
The designers are looking for bettern pattern images.